+++ src/Constants/Config.php
@@ -24,12 +24,13 @@ class Config
     // ====================
     // Checkout Config Keys
     // ====================
-    public const CHECKOUT_CONFIG_KEY = 'checkout';
-    public const CHECKOUT_LIVE_MODE = self::CHECKOUT_CONFIG_KEY . 'LiveMode';
-    public const CHECKOUT_USERNAME = self::CHECKOUT_CONFIG_KEY . 'MerchantUsername';
-    public const CHECKOUT_PASSWORD = self::CHECKOUT_CONFIG_KEY . 'MerchantPassword';
+    public const CHECKOUT_CONFIG_KEY        = 'checkout';
+    public const CHECKOUT_LIVE_MODE         = self::CHECKOUT_CONFIG_KEY . 'LiveMode';
+    public const CHECKOUT_USERNAME          = self::CHECKOUT_CONFIG_KEY . 'MerchantUsername';
+    public const CHECKOUT_PASSWORD          = self::CHECKOUT_CONFIG_KEY . 'MerchantPassword';
     public const CHECKOUT_TRANSACTION_TYPES = self::CHECKOUT_CONFIG_KEY . 'TransactionTypes';
-    public const CHECKOUT_LANGUAGE = self::CHECKOUT_CONFIG_KEY . 'Language';
+    public const CHECKOUT_LANGUAGE          = self::CHECKOUT_CONFIG_KEY . 'Language';
+    public const CHECKOUT_TOKENIZATION      = self::CHECKOUT_CONFIG_KEY . 'Tokenization';
 
     /**
      * Nested Transaction Types suffix
@@ -69,6 +70,7 @@ class Config
             self::CHECKOUT_PASSWORD,
             self::CHECKOUT_TRANSACTION_TYPES,
             self::CHECKOUT_LANGUAGE,
+            self::CHECKOUT_TOKENIZATION
         ];
     }
 }