+++ src/Core/Payment/Tokenization/TokenizationConsumerDefinition.php
@@ -1,0 +1,58 @@
+<?php declare(strict_types=1);
+/*
+ * Copyright (C) 2021 emerchantpay Ltd.
+ *
+ * This program is free software; you can redistribute it and/or
+ * modify it under the terms of the GNU General Public License
+ * as published by the Free Software Foundation; either version 2
+ * of the License, or (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * @author      emerchantpay
+ * @copyright   2021 emerchantpay Ltd.
+ * @license     http://opensource.org/licenses/gpl-2.0.php GNU General Public License, version 2 (GPL-2.0)
+ */
+
+namespace Emerchantpay\Genesis\Core\Payment\Tokenization;
+
+use Shopware\Core\Framework\DataAbstractionLayer\EntityDefinition;
+use Shopware\Core\Framework\DataAbstractionLayer\Field\Flag\PrimaryKey;
+use Shopware\Core\Framework\DataAbstractionLayer\Field\Flag\Required;
+use Shopware\Core\Framework\DataAbstractionLayer\Field\IdField;
+use Shopware\Core\Framework\DataAbstractionLayer\Field\StringField;
+use Shopware\Core\Framework\DataAbstractionLayer\FieldCollection;
+
+class TokenizationConsumerDefinition extends EntityDefinition
+{
+    public const ENTITY_NAME = 'emerchantpay_tokenization_consumer';
+    public const MAX_EMAIL_LENGTH = 320;
+    public const MAX_CONSUMER_ID_LENGTH = 64;
+
+    public function getEntityName(): string
+    {
+        return self::ENTITY_NAME;
+    }
+
+    public function getEntityClass(): string
+    {
+        return TokenizationConsumerEntity::class;
+    }
+
+    public function getCollectionClass(): string
+    {
+        return TokenizationConsumerCollection::class;
+    }
+
+    protected function defineFields(): FieldCollection
+    {
+        return new FieldCollection([
+            (new IdField('id', 'id'))->addFlags(new Required(), new PrimaryKey()),
+            (new StringField('email', 'email', self::MAX_EMAIL_LENGTH)),
+            (new StringField('consumer_id', 'consumerId', self::MAX_CONSUMER_ID_LENGTH))
+        ]);
+    }
+}