+++ src/Core/Payment/Tokenization/TokenizationConsumerEntity.php
@@ -1,0 +1,63 @@
+<?php declare(strict_types=1);
+/*
+ * Copyright (C) 2021 emerchantpay Ltd.
+ *
+ * This program is free software; you can redistribute it and/or
+ * modify it under the terms of the GNU General Public License
+ * as published by the Free Software Foundation; either version 2
+ * of the License, or (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * @author      emerchantpay
+ * @copyright   2021 emerchantpay Ltd.
+ * @license     http://opensource.org/licenses/gpl-2.0.php GNU General Public License, version 2 (GPL-2.0)
+ */
+
+namespace Emerchantpay\Genesis\Core\Payment\Tokenization;
+
+use Shopware\Core\Framework\DataAbstractionLayer\Entity;
+use Shopware\Core\Framework\DataAbstractionLayer\EntityIdTrait;
+
+class TokenizationConsumerEntity extends Entity
+{
+    use EntityIdTrait;
+
+    protected $email;
+    protected $consumerId;
+
+    /**
+     * @return string
+     */
+    public function getConsumerId(): string
+    {
+        return $this->consumerId;
+    }
+
+    /**
+     * @param string $consumerId
+     */
+    public function setConsumerId(string $consumerId): void
+    {
+        $this->consumerId = $consumerId;
+    }
+
+    /**
+     * @return string
+     */
+    public function getEmail(): string
+    {
+        return $this->email;
+    }
+
+    /**
+     * @param string $email
+     */
+    public function setEmail(string $email): void
+    {
+        $this->email = $email;
+    }
+}