+++ CHANGELOG.md
@@ -1,7 +1,14 @@
 Change Log
 ---------------------
 
+__1.1.0__
+-----
+* Added config option for enabling Web Payment Form Tokenization service. Please, contact your account manager before enabling tokenization (#839)
+* Updated to work with Shopware v. 6.4.10 (#859)
+* Fixed payment_id database migration causing issues with the older Doctrine versions (#862)
+
 __1.0.4__
+-----
 * Updated Genesis PHP SDK library to version 1.20.1 (#809)
 * Added new transaction type Apple Pay via Web Payment Form with support of its methods (#810):
   * Authorize